requirements.txt
Django>=3.1.3,<3.2.*
psycopg2>=2.8.6,<2.9.*

graphene-django>=2.13.0,<2.14.*

django-graphql-jwt>=0.3.1,<0.4.*